+++ NEWS.md
@@ -1,6 +1,14 @@
 
 ## reticulate 1.19 (UNRELEASED)
 
+- `virtualenv_create()` gains the `pip_version` and `setuptools_version`
+  arguments, allowing users to control the versions of `pip` and `setuptools`
+  used when initializing the virtual environment. The `extra` argument can
+  also now be used to pass arbitrary command line arguments when necessary.
+
+- `virtualenv_create()` gains the `module` argument, used to control whether
+  `virtualenv` or `venv` is used to create the requested virtual environment.
+  
 - `py_to_r.datetime.datetime` no longer errs when `tzname` is `NULL`, and
   instead assumes the time is formated for `UTC`. (#876)